melodyMingle/playlists/serializers.py
from rest_framework import serializers
from .models import Track

class TrackSerializer(serializers.ModelSerializer):
    class Meta:
        model = Track
        fields = ['id', 'title', 'artist', 'genre',
                  'track_file_url', 'cover_image_url', 'duration']