melodyMingle/requirements.txt
Django==4.2.7
Pillow==10.1.0
PyJWT==2.8.0
asgiref==3.7.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
pip==23.3.1
pytz==2023.3.post1
setuptools==57.0.0
sqlparse==0.4.4
tzdata==2023.3
wheel==0.36.2
gunicorn==21.2.0